Adding Builds Folder
TestRpg_0.1.apk is before I changed the idea from a dungeon crawler, you'll start on Floor 1 in the dungeon. 
TestRpg0.2 is the refactored version to be more "card-based." No more dungeon crawling. You'll start in the overworld.

None of these builds have tutorials, they are pretty crude and unpolished, but remember that this is all a WIP. I'm currently building the base mechanics for the game, the actual "gameplay" included in the builds are simply basic areas quickly built to test the mechanics.

Regardless, try to have some 'fun' with these!